Rename: /api/auth/auth.js -> userApi.js
